Added Files, scrapped code, note to self
Added files:
Test File (meant to do testing in)
ORE_FILE (place to store scrapped code)

Started over today and put old code in de ORE_FILE in order to preserve the knowledge and also to borrow from it and learn from it. If stuck take a look in that code.

Most of the code is starting to work... Still some ways off.

note to self for next time

// For next time

Take a look at why the code will PAUSE the movement of the bullseye even when not clicking inside the box.

Start on solving the problem of clicking INSIDE the BULLSEYE and then having SOMETHING HAPPEN. Start off with it just flashing something in the screen. 

Next step would be to clean up the code. EG write boolean functions instead of void functions where possible.

After that write SCORE COUNTER

TERNARY operator still doesn't work as it should. Take a look at this while cleaning up the code, but don't spend too much time on it please. it looks pretty, but it can wait for now.... perhaps the penny will drop sometime soon.